Move multiple classes to common
Includes:
- DynamoConfig
- DynamoClientModule
- DynamoConfigModule
- SNSConfig
- SNSMessage
- SNSConfigModule
- SNSClientModule
- SQSConfigModule
- SQSClientModule
