+++ common/src/main/scala/uk/ac/wellcome/finatra/modules/SNSClientModule.scala
@@ -1,16 +1,12 @@
-package uk.ac.wellcome.platform.finatra.modules
+package uk.ac.wellcome.finatra.modules
 
 import javax.inject.Singleton
 
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
+import uk.ac.wellcome.models.SNSConfig
 
 import com.amazonaws.services.sns._
-
-/** Sent to request a change in the desired count of an ECS service. */
-case class ECSServiceScheduleRequest(cluster: String,
-                                     service: String,
-                                     desired_count: Long)
 
 object SNSClientModule extends TwitterModule {
   override val modules = Seq(SNSConfigModule)
