+++ common/src/main/scala/uk/ac/wellcome/finatra/modules/SNSConfigModule.scala
@@ -1,11 +1,11 @@
-package uk.ac.wellcome.platform.finatra.modules
+package uk.ac.wellcome.finatra.modules
 
 import javax.inject.Singleton
 
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
+import uk.ac.wellcome.models.SNSConfig
 
-case class SNSConfig(region: String, topicArn: String)
 
 object SNSConfigModule extends TwitterModule {
   private val region = flag[String]("aws.region", "eu-west-1", "AWS region")
