+++ common/src/main/scala/uk/ac/wellcome/finatra/modules/SQSConfigModule.scala
@@ -1,11 +1,10 @@
-package uk.ac.wellcome.platform.ingestor.modules
+package uk.ac.wellcome.finatra.modules
 
 import javax.inject.Singleton
 
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
-
-case class SQSConfig(region: String, queueUrl: String)
+import uk.ac.wellcome.models.SQSConfig
 
 object SQSConfigModule extends TwitterModule {
   private val region = flag[String]("aws.region", "eu-west-1", "AWS region")
