+++ common/src/main/scala/uk/ac/wellcome/models/DynamoConfig.scala
@@ -1,0 +1,6 @@
+package uk.ac.wellcome.models
+
+case class DynamoConfig(region: String,
+                        applicationName: String,
+                        arn: String,
+                        table: String)
