+++ common/src/main/scala/uk/ac/wellcome/models/ECSServiceScheduleRequest.scala
@@ -1,0 +1,5 @@
+package uk.ac.wellcome.models
+
+case class ECSServiceScheduleRequest(cluster: String,
+                                     service: String,
+                                     desired_count: Long)
