+++ common/src/main/scala/uk/ac/wellcome/models/SNSConfig.scala
@@ -1,0 +1,3 @@
+package uk.ac.wellcome.models
+
+case class SNSConfig(region: String, topicArn: String)
