+++ common/src/main/scala/uk/ac/wellcome/models/SQSConfig.scala
@@ -1,0 +1,3 @@
+package uk.ac.wellcome.models
+
+case class SQSConfig(region: String, queueUrl: String)
