+++ transformer/src/main/scala/uk/ac/wellcome/transformer/modules/KinesisWorker.scala
@@ -12,9 +12,12 @@ import com.amazonaws.services.dynamodbv2._
 import com.amazonaws.services.dynamodbv2.streamsadapter.AmazonDynamoDBStreamsAdapterClient
 import com.amazonaws.services.kinesis.clientlibrary.lib.worker._
 import com.twitter.inject.{Injector, Logging, TwitterModule}
+import uk.ac.wellcome.models.DynamoConfig
 
-import uk.ac.wellcome.platform.finatra.modules._
-import uk.ac.wellcome.finatra.modules.AkkaModule
+import uk.ac.wellcome.finatra.modules.{
+  AkkaModule,
+  DynamoConfigModule
+}
 
 object KinesisWorker extends TwitterModule {
   override val modules = Seq(StreamsRecordProcessorFactoryModule,
