+++ transformer/src/main/scala/uk/ac/wellcome/transformer/modules/StreamRecordProcessorFactoryModule.scala
@@ -12,7 +12,7 @@ import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
 
 import akka.actor.ActorRef
-import uk.ac.wellcome.platform.finatra.modules._
+import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.platform.transformer.modules._
 import uk.ac.wellcome.models.ActorRegister
